% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demuxHashTags.R
\name{demuxHashTags}
\alias{demuxHashTags}
\title{a simple HTO demuxing function based on hashedDrops (support genotype later)}
\usage{
demuxHashTags(sce, hto = NULL, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object with a colData column \code{keep}}

\item{hto}{a character vector of HTO names}

\item{...}{additional arguments to be passed to \code{\link{hashedDrops}}}
}
\value{
a SingleCellExperiment object with new colData columns (see Details)
}
\description{
DropletUtils::hashedDrops assumes empty droplets have been removed. Schiele
assumes that sce will have colData column \code{keep} to indicate which cells,
and \code{keep} column will be a logical variable. To enforce these assumptions,
we refuse to attempt hash-tag oligo demuxing unless assumptions are met.
}
\details{
The output of hashedDrops is placed in metadata(sce)$hashed, and three
columns ("HTO", "Confident", "Doublet") are added to colData(sce).
For columns where \code{keep} is FALSE, these values are set to NA.
A note is also placed in metadata(sce) regarding the status of cells.
}
\seealso{
\code{\link{hashedDrops}}
}
