% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMultiple.R
\name{findMultiple}
\alias{findMultiple}
\title{see if a package is installed more than once}
\usage{
findMultiple(pkg = "Seurat")
}
\arguments{
\item{pkg}{name of the package}
}
\value{
a data.frame with columns 'Package' and 'LibPath' of 0 or more rows
}
\description{
see if a package is installed more than once
}
\examples{
findMultiple("Seurat")

}
