% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExp.R
\name{getExp}
\alias{getExp}
\title{generic mainExp/altExp getter}
\usage{
getExp(sce, e, a = "counts")
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{e}{name of the experiment to get}

\item{a}{name of the assay to get ("counts")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ the contents of the requested assay for the requested experiment
}\if{html}{\out{</div>}}
}
\description{
If a named experiment isn't in the altExps of a SingleCellExperiment,
get the requested assay from the mainExp, iff mainExpName(sce) == e.
}
\seealso{
HTOs

ADTs

mRNA
}
