% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOs.R
\name{HTOs}
\alias{HTOs}
\title{accessor for the 'HTO' altExp in a sensibly processed SingleCellExperiment}
\usage{
HTOs(sce, a = "counts")
}
\arguments{
\item{sce}{a SingleCellExperiment, ideally one from sceH5 or sceMTX}
}
\value{
a matrix of HTO counts, with rownames corresponding to the HTO names
}
\description{
accessor for the 'HTO' altExp in a sensibly processed SingleCellExperiment
}
\seealso{
\code{\link{demuxHashTags}}

\code{\link{sceMTX}}

\code{\link{sceH5}}

\code{\link{pHTO}}
}
