% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kneePlot.R
\name{kneePlot}
\alias{kneePlot}
\title{Create a halfway decent knee plot for shallow CITEseq runs}
\usage{
kneePlot(sce, target = 10000)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{target}{The target number of cells (10000)}
}
\value{
A ggplot2 object
}
\description{
Create a halfway decent knee plot for shallow CITEseq runs
}
