% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mtxs.R
\name{load_mtxs}
\alias{load_mtxs}
\title{Load a bunch of MatrixMarket triples (features, cells, matrix) into an SCE.}
\usage{
load_mtxs(runs, ..., BPPARAM = SerialParam())
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, try using the entries)}

\item{...}{arguments to pass on to load_mtx()}

\item{BPPARAM}{if running in parallel, provide a MulticoreParam or similar}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a SingleCellExperiment
}\if{html}{\out{</div>}}
}
\description{
Load a bunch of MatrixMarket triples (features, cells, matrix) into an SCE.
}
