% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBusText.R
\name{loadBusText}
\alias{loadBusText}
\title{Read a .bus.txt file produced by \verb{bustools text}}
\usage{
loadBusText(file)
}
\arguments{
\item{file}{the bustools text file to import}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a data.frame 
}\if{html}{\out{</div>}}
}
\description{
This is rarely done in practice (BUS is more efficient) but useful for demos.
}
\details{
This function works fine on gzipped files too.
}
\examples{
bus_txt <- system.file("extdata", "short.bus.txt.gz", package="Schiele")
bus <- loadBusText(bus_txt)

# how many fragments per cell? (low-pass CITEseq example)
summary(as.integer(with(bus, rowsum(count, barcode))))

# how many fragments per equivalence class (feature)?
summary(as.integer(with(bus, rowsum(count, set))))

# how many equivalence classes (features) per cell?
summary(sapply(split(bus, bus$barcode), function(x) length(unique(x$set))))

}
