% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMatrix.R
\name{loadMatrix}
\alias{loadMatrix}
\title{Load features, cells, feature x cell counts from files, as from 10X}
\usage{
loadMatrix(path = ".", verbose = TRUE, frags = NULL, splt = FALSE, spcol = 3)
}
\arguments{
\item{path}{where the counts files are (features, barcodes, mtx) (".")}

\item{verbose}{be verbose? (TRUE)}

\item{frags}{optional patterns to scan for features, barcodes, matrix}

\item{splt}{split by feature type? (FALSE)}

\item{spcol}{if splitting by feature type, which column to use? (3)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a dgCMatrix 
}\if{html}{\out{</div>}}
}
\description{
The returned matrix is a dgCMatrix as used by (e.g.) SingleCellExperiment.
This function is useful when processing data from (e.g.) kallisto|bustools,
for example, CITE-seq data processed with KITE, or when comparing uniquely
mapped vs. multimapped counts, as one might do for repetitive elements.
}
\details{
This function works fine on gzipped files (tested on a 180K x 2761K matrix).
}
