% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logADTs.R
\name{logADTs}
\alias{logADTs}
\title{logcounts for the 'ADT' altExp in a sensibly processed SingleCellExperiment}
\usage{
logADTs(sce)
}
\arguments{
\item{sce}{a SingleCellExperiment, ideally from sceMTX or sceH5}
}
\value{
a matrix of ADT logcounts, rownames corresponding to the ADT names
}
\description{
logcounts for the 'ADT' altExp in a sensibly processed SingleCellExperiment
}
\seealso{
\code{\link{sceH5}}
}
