% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markKept.R
\name{markKept}
\alias{markKept}
\title{Attempt to call nonempty cells based on a target cell number from a kneeplot}
\usage{
markKept(sce, targetCells = 20000, targetUMIs = 50, column = "UMIs", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{targetCells}{the number of cells supposedly loaded (20000)}

\item{targetUMIs}{minimum UMIs to aim for if UMIs\link{targetCells} is low (50)}

\item{column}{optional colData name for cached results ("UMIs")}

\item{...}{additional arguments to perhaps pass to emptyDrops}
}
\value{
a SingleCellExperiment object with a new column \code{keep}
}
\description{
Sometimes DropletUtils::emptyDrops() will fail, especially on shallow runs
designed to determine if a CITEseq experiment is worth sequencing deeply.
This can occur even when the shallow libraries are quite clean. Therefore,
we provide this wrapper function to ease the pain of QC'ing shallow runs.
}
\seealso{
kneePlot

emptyDrops
}
