% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA.R
\name{mRNA}
\alias{mRNA}
\title{accessor for the 'RNA' altExp in a sensibly processed SingleCellExperiment}
\usage{
mRNA(sce, a = "logcounts")
}
\arguments{
\item{sce}{a SingleCellExperiment, ideally one from \code{\link{SCE}}}
}
\value{
a matrix of RNA abundance, with rownames corresponding to genes
}
\description{
Note that unlike ADTs and HTOs, the default assay is 'logcounts' for \code{mRNA}.
}
\seealso{
\code{\link{SCE}}
}
