% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oink.R
\name{oink}
\alias{oink}
\title{Load some perfectly good data into a Seurat object.}
\usage{
oink(runs, ...)
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, try using the entries)}

\item{...}{arguments to pass on to load_mtxs()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a Seurat object, if Seurat is (tragically) loaded
}\if{html}{\out{</div>}}
}
\description{
Load some perfectly good data into a Seurat object.
}
