% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHTO.R
\name{pHTO}
\alias{pHTO}
\title{moronically simple function to calculate Pr(HTO=hto|counts)}
\usage{
pHTO(sce, off = 0.1)
}
\arguments{
\item{sce}{a SingleCellExperiment object with !is.null(altExp(sce, "HTO"))}

\item{off}{offtarget probability (default is 0.1 total for all !whichHTO)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ same sce but with new column "pHTO" (and "HTO" if not present)
}\if{html}{\out{</div>}}
}
\description{
Use a multinomial test to calculate HTO probability.
}
