% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSparseMat.R
\name{readSparseMat}
\alias{readSparseMat}
\title{read a MatrixMarket file and return a dgCMatrix with features as rows}
\usage{
readSparseMat(mat)
}
\arguments{
\item{mat}{path to a MatrixMarket file}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a dgCMatrix
}\if{html}{\out{</div>}}
}
\description{
This is a wrapper around \code{\link{Matrix::readMM}} that returns a
\code{\link{dgCMatrix}} with features as rows. It is factored out solely
to isolate any fallout from changes to the \code{\link{Matrix}} package.
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  Matrix::readMM
}\if{html}{\out{</div>}}
}
