% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{sce}
\alias{sce}
\title{Load multiple count matrices into a SingleCellExperiment}
\usage{
sce(runs, verbose = FALSE, ..., BPPARAM = SerialParam())
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, try using the entries)}

\item{verbose}{be verbose? (FALSE)}

\item{...}{other arguments to pass on to load_mtx()}

\item{BPPARAM}{if running in parallel, provide a MulticoreParam or similar}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a SingleCellExperiment
}\if{html}{\out{</div>}}
}
\description{
Load multiple count matrices into a SingleCellExperiment
}
