% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceH5.R
\name{sceH5}
\alias{sceH5}
\title{Read a 10X-formatted .h5 file into a SummarizedExperiment}
\usage{
sceH5(path, splt = FALSE, spcol = 3, verbose = TRUE)
}
\arguments{
\item{path}{path to the .h5 file}

\item{splt}{split by feature type, e.g. HTO vs. ADT? (FALSE)}

\item{spcol}{if splitting by feature type, which column to use? (3)}

\item{verbose}{be verbose? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       a SingleCellExperiment
}\if{html}{\out{</div>}}
}
\description{
This function is useful for data processed by (e.g.) cellbender.
}
