% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceMTX.R
\name{sceMTX}
\alias{sceMTX}
\title{Load multiple count matrices into a SingleCellExperiment}
\usage{
sceMTX(runs, splt = TRUE, verbose = FALSE, ..., BPPARAM = NULL)
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, try using the entries)}

\item{splt}{attempt to split HTOs from ADTs/mRNA? (only if found)}

\item{verbose}{be verbose? (FALSE)}

\item{...}{other arguments to pass on to loadMatrix()}

\item{BPPARAM}{if running in parallel, provide a MulticoreParam or similar}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a SingleCellExperiment
}\if{html}{\out{</div>}}
}
\description{
Load multiple count matrices into a SingleCellExperiment
}
