% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{ADTs}
\alias{ADTs}
\title{accessor for the 'ADT' altExp in a sensibly processed SingleCellExperiment}
\usage{
ADTs(sce, a = "counts")
}
\arguments{
\item{sce}{a SingleCellExperiment, ideally one from \code{\link{sce}}}
}
\value{
a matrix of ADT counts, with rownames corresponding to the ADT names
}
\description{
accessor for the 'ADT' altExp in a sensibly processed SingleCellExperiment
}
\seealso{
\code{\link{sce}}
}
