% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oink.R
\name{oink}
\alias{oink}
\title{Load multiple count matrices into a Seurat object (perhaps)}
\usage{
oink(runs, ...)
}
\arguments{
\item{runs}{a named vector of paths (if unnamed, try using the entries)}

\item{...}{arguments to pass on to sce()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   a Seurat object, if Seurat is (tragically) loaded, else an SCE
}\if{html}{\out{</div>}}
}
\description{
Load multiple count matrices into a Seurat object (perhaps)
}
